requirements.txt
psutil
requests
pyjwt
watchdog
